+++ client/src/index.js
@@ -14,27 +14,30 @@ ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
       <Routes>
-          <Route path="/" element={<App />} />
-          <Route path="register" element={<Register actionLabel={`Register`}/>} />
-          <Route path="dashboard/:companyId" element={<DashBoard />} />
-          <Route path="dashboard/:companyId/jnf" element={<Jnf />} />
-          <Route path="dashboard/:companyId/inf" element={<Inf actionLabel={`Submit`}/>} />
-          <Route
-            path="dashboard/:companyId/preview/inf"
-            element={<PreviewInf />}
-          />
-          <Route
-            path="dashboard/:companyId/preview/jnf"
-            element={<PreviewJnf />}
-          />
-          <Route
-            path="dashboard/:companyId/updateInf"
-            element={<Inf actionLabel={`Update`} />}
-          />
-          <Route
-            path="dashboard/:companyId/updateCompany"
-            element={<Register actionLabel={`Update`} />}
-          />
+        <Route path="/" element={<App />} />
+        <Route
+          path="register"
+          element={<Register actionLabel={`Register`} />}
+        />
+        <Route path="dashboard/:companyId" element={<DashBoard />} />
+        <Route path="dashboard/:companyId/jnf" element={<Jnf />} />
+        <Route path="dashboard/:companyId/inf" element={<Inf></Inf>} />
+        <Route
+          path="dashboard/:companyId/preview/inf"
+          element={<PreviewInf />}
+        />
+        <Route
+          path="dashboard/:companyId/preview/jnf"
+          element={<PreviewJnf />}
+        />
+        <Route
+          path="dashboard/:companyId/updateInf"
+          element={<Inf actionLabel={`Update`} />}
+        />
+        <Route
+          path="dashboard/:companyId/updateCompany"
+          element={<Register actionLabel={`Update`} />}
+        />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>,
